test_case/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
idna==3.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12
